Driver and 4 & 5
driver added. need to fix when using the wrong username and password

function 5 seems to work fine.
function 4 compiles and runs but doesn’t display names for some reason.
think it has something to do with getting results from query but the
sql statement should be correct.

also a change to establish friendship where it inserts a flipped
friendship so when looking for friendships it is easier to select